pyproject.toml
[project]
name = "react-agent-implementation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "openai>=1.93.0",
    "pandas>=2.3.1",
    "torch>=2.7.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.0",
    "underthesea>=6.8.4",
    "transformers>=4.53.2",
    "accelerate>=1.9.0",
    "pillow>=11.3.0",
    "tavily-python>=0.7.9",
    "datasets>=4.0.0",
    "trl>=0.19.1",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "uv>=0.8.0",
]
